fixedf SetMaximumUnknownFramesToSkip

+++ src/Serum.h
@@ -16,8 +16,8 @@ class Serum
   void SetStandardPalette(const uint8_t* palette, const int bitDepth);
   bool ColorizeWithMetadata(uint8_t* frame, int width, int height, uint8_t* palette, uint8_t* rotations,
                             uint32_t* triggerID, uint32_t* hashcode, int* frameID);
-  void SetIgnoreUnknownFramesTimeout(uint8_t milliseconds);
-  void SetMaximumUnknownFramesToSkip(uint8_t maximum);
+  void SetIgnoreUnknownFramesTimeout(int milliseconds);
+  void SetMaximumUnknownFramesToSkip(int maximum);
 
  private:
   Serum(int width, int height);
